+++ client/src/assets/img/dice/dice1.PNG
