+++ client/src/assets/img/dice/dice6.PNG
