+++ client/src/components/organisms/Dice.tsx
@@ -1,15 +1,15 @@
 import React, { useState, useEffect } from 'react';
-import diceState from '@store/Dice_number';
+import diceState from '@store/dice';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import styled from 'styled-components';
-import dice_1 from '@img/dice/dice_1.PNG';
-import dice_2 from '@img/dice/dice_2.PNG';
-import dice_3 from '@img/dice/dice_3.PNG';
-import dice_4 from '@img/dice/dice_4.PNG';
-import dice_5 from '@img/dice/dice_5.PNG';
-import dice_6 from '@img/dice/dice_6.PNG';
+import dice1 from '@img/dice/dice1.PNG';
+import dice2 from '@img/dice/dice2.PNG';
+import dice3 from '@img/dice/dice3.PNG';
+import dice4 from '@img/dice/dice4.PNG';
+import dice5 from '@img/dice/dice5.PNG';
+import dice6 from '@img/dice/dice6.PNG';
 
-const dice_list = [dice_1, dice_2, dice_3, dice_4, dice_5, dice_6];
+const dice_list = [dice1, dice2, dice3, dice4, dice5, dice6];
 const Dice_svg = styled.svg`
   z-index: 1;
   position: absolute;
